added product repo and product service in different packages
