+++ src/main/java/com/example/service/ProductService.java
@@ -1,30 +1,32 @@
 package com.example.service;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
+
+import com.example.repository.ProductRepository;
 import com.example.webSimple.entities.Product;
 
 @Service
 public class ProductService {
-	private static List<Product> products=new ArrayList<>();
 
-	public static List<Product> getAllProducts() {
-		return products;
+	public List<Product> getAllProducts() {
+		return ProductRepository.getAllProducts();
 	}
 
-	public static Product getProductById(int id) {
-		for (Product product : products) {
-			if (product.getId() == id) {
-				return product;
-			}
-		}
-		return null;
+	public Product getProductById(int id) {
+		return ProductRepository.getProductById(id);
 	}
 
-	public static void addProduct(Product product) {
-		products.add(product);
+	public void addProduct(Product product) {
+		ProductRepository.addProduct(product);
 	}
-
+	
+	public void deleteProduct(Product product) {
+		ProductRepository.removeProduct(product);
+	}
+	
+	public void deleteProductById(int id) {
+		ProductRepository.removeProduct(id);
+	}
 }
